refactor: Update Supabase page to list users from user_table
This commit refactors the Supabase page to list users from the user_table in the Supabase database. The previous implementation was fetching all addresses, but now it fetches the "content" field from the user_table. This change improves the functionality of the page and ensures that the correct data is displayed to the users.