refactor: Add user admin check to UserAdmin page
This commit refactors the UserAdmin page to include a check for user admin status. It retrieves the user's information from the Clerk API and fetches the corresponding user from the User table in the Prisma database. If the user is an admin, the page displays "Admin Page", otherwise it displays "Not an Admin". This change improves the functionality of the UserAdmin page by distinguishing between regular users and admin users.