refactor: Update UserAdmin page to display admin status
This commit updates the UserAdmin page to include the user's admin status. It retrieves the user's information from the Clerk API and fetches the corresponding user from the User table in the Prisma database. If the user is an admin, the page displays "Admin Page", otherwise it displays "Not an Admin". This change improves the functionality of the UserAdmin page by providing clear information about the user's admin status.